pom dependencies upgraded
