--- grpc-server/src/main/java/com/github/anjeyy/infrastructure/exception/handler/GrpcServiceAdviceIsPresent.java
@@ -1,49 +1,0 @@
-/*
- * Copyright (c) 2016-2020 Michael Zhang <yidongnan@gmail.com>
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
- * documentation files (the "Software"), to deal in the Software without restriction, including without limitation the
- * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the
- * Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE
- * WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
- * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package com.github.anjeyy.infrastructure.exception.handler;
-
-import com.github.anjeyy.infrastructure.annotation.GrpcAdvice;
-import java.util.Objects;
-import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
-import org.springframework.context.annotation.ConditionContext;
-import org.springframework.context.annotation.ConfigurationCondition;
-import org.springframework.core.type.AnnotatedTypeMetadata;
-
-/**
- * Condition to check if {@link GrpcAdvice @GrpcServiceAdvice} is present. Mainly checking if
- * {@link GrpcAdviceDiscoverer} should be a instantiated.<br>
- * <br>
- *
- * @author Andjelko Perisic (andjelko.perisic@gmail.com)
- * @see GrpcAdviceDiscoverer
- */
-public class GrpcServiceAdviceIsPresent implements ConfigurationCondition {
-
-    @Override
-    public ConfigurationPhase getConfigurationPhase() {
-        return ConfigurationPhase.REGISTER_BEAN;
-    }
-
-    @Override
-    public boolean matches(ConditionContext context, AnnotatedTypeMetadata metadata) {
-
-        ConfigurableListableBeanFactory safeBeanFactory =
-            Objects.requireNonNull(context.getBeanFactory(), "ConfigurableListableBeanFactory is null");
-        return !safeBeanFactory.getBeansWithAnnotation(GrpcAdvice.class).isEmpty();
-    }
-}
